+++ Changelog.md
@@ -1,12 +1,12 @@
 # Changelog v1.4.1
 
-## Additions
-
-
 ## Changes
 - Changed textures for logs/stems
 - Changed textures for bamboo & stripped bamboo block
 - Changed texture for stripped logs/stems
+
+## Breaking Changes
+- Removed stripped mossy wood
 
 ### Downloads
 [CurseForge](https://curseforge.com/minecraft/mc-mods/nemos-mossy-blocks) <br>